requirements.txt
feedparser>=5.2
python-twitter>=3.2
